Ensure XPSR register is exposed by regenerating bindings on build.
Unicorn_mode requires the XPSR register to be exposed and applies
patches to this effect. However, it does not regenerate the
python language bindings, meaning that the newly exposed registers are
unavailable to hal-fuzz. This patch fixes this.

Also included: some updated .gitignore for all binaries that should
not be commited. native_hooks.so is also removed as a binary.